requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
qrcode==7.4.2
pycryptodome==3.19.0
cryptography==41.0.7
pillow-simd==10.0.0
